+++ homeassistant/components/mqtt/device_tracker/schema_discovery.py
@@ -25,17 +25,20 @@ from homeassistant.helpers.dispatcher import (
     async_dispatcher_send,
 )
 
-from .. import (
-    MqttAttributes,
-    MqttAvailability,
-    MqttDiscoveryUpdate,
-    MqttEntityDeviceInfo,
-    subscription,
-)
+from .. import subscription
 from ... import mqtt
 from ..const import ATTR_DISCOVERY_HASH, CONF_QOS, CONF_STATE_TOPIC
 from ..debug_info import log_messages
 from ..discovery import MQTT_DISCOVERY_DONE, MQTT_DISCOVERY_NEW, clear_discovery_hash
+from ..mixins import (
+    MQTT_AVAILABILITY_SCHEMA,
+    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+    MQTT_JSON_ATTRS_SCHEMA,
+    MqttAttributes,
+    MqttAvailability,
+    MqttDiscoveryUpdate,
+    MqttEntityDeviceInfo,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -46,7 +49,7 @@ CONF_SOURCE_TYPE = "source_type"
 PLATFORM_SCHEMA_DISCOVERY = (
     mqtt.MQTT_RO_PLATFORM_SCHEMA.extend(
         {
-            vol.Optional(CONF_DEVICE): mqtt.MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+            vol.Optional(CONF_DEVICE): MQTT_ENTITY_DEVICE_INFO_SCHEMA,
             vol.Optional(CONF_ICON): cv.icon,
             vol.Optional(CONF_NAME): cv.string,
             vol.Optional(CONF_PAYLOAD_HOME, default=STATE_HOME): cv.string,
@@ -55,8 +58,8 @@ PLATFORM_SCHEMA_DISCOVERY = (
             vol.Optional(CONF_UNIQUE_ID): cv.string,
         }
     )
-    .extend(mqtt.MQTT_AVAILABILITY_SCHEMA.schema)
-    .extend(mqtt.MQTT_JSON_ATTRS_SCHEMA.schema)
+    .extend(MQTT_AVAILABILITY_SCHEMA.schema)
+    .extend(MQTT_JSON_ATTRS_SCHEMA.schema)
 )
