+++ homeassistant/components/mqtt/tag.py
@@ -3,7 +3,7 @@ import logging
 
 import voluptuous as vol
 
-from homeassistant.const import CONF_PLATFORM, CONF_VALUE_TEMPLATE
+from homeassistant.const import CONF_DEVICE, CONF_PLATFORM, CONF_VALUE_TEMPLATE
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.device_registry import EVENT_DEVICE_REGISTRY_UPDATED
 from homeassistant.helpers.dispatcher import (
@@ -11,24 +11,22 @@ from homeassistant.helpers.dispatcher import (
     async_dispatcher_send,
 )
 
-from . import (
-    ATTR_DISCOVERY_HASH,
-    ATTR_DISCOVERY_TOPIC,
-    CONF_CONNECTIONS,
-    CONF_DEVICE,
-    CONF_IDENTIFIERS,
-    CONF_QOS,
-    CONF_TOPIC,
-    DOMAIN,
-    cleanup_device_registry,
-    subscription,
-)
+from . import CONF_QOS, CONF_TOPIC, DOMAIN, subscription
 from .. import mqtt
+from .const import ATTR_DISCOVERY_HASH, ATTR_DISCOVERY_TOPIC
 from .discovery import (
     MQTT_DISCOVERY_DONE,
     MQTT_DISCOVERY_NEW,
     MQTT_DISCOVERY_UPDATED,
     clear_discovery_hash,
+)
+from .mixins import (
+    CONF_CONNECTIONS,
+    CONF_IDENTIFIERS,
+    MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+    cleanup_device_registry,
+    device_info_from_config,
+    validate_device_has_at_least_one_identifier,
 )
 from .util import valid_subscribe_topic
 
@@ -39,12 +37,12 @@ TAGS = "mqtt_tags"
 
 PLATFORM_SCHEMA = mqtt.MQTT_BASE_PLATFORM_SCHEMA.extend(
     {
-        vol.Optional(CONF_DEVICE): mqtt.MQTT_ENTITY_DEVICE_INFO_SCHEMA,
+        vol.Optional(CONF_DEVICE): MQTT_ENTITY_DEVICE_INFO_SCHEMA,
         vol.Optional(CONF_PLATFORM): "mqtt",
         vol.Required(CONF_TOPIC): valid_subscribe_topic,
         vol.Optional(CONF_VALUE_TEMPLATE): cv.template,
     },
-    mqtt.validate_device_has_at_least_one_identifier,
+    validate_device_has_at_least_one_identifier,
 )
 
 
@@ -236,7 +234,7 @@ async def _update_device(hass, config_entry, config):
     """Update device registry."""
     device_registry = await hass.helpers.device_registry.async_get_registry()
     config_entry_id = config_entry.entry_id
-    device_info = mqtt.device_info_from_config(config[CONF_DEVICE])
+    device_info = device_info_from_config(config[CONF_DEVICE])
 
     if config_entry_id is not None and device_info is not None:
         device_info["config_entry_id"] = config_entry_id
